+++ client/src/components/BackButton.tsx
@@ -13,7 +13,7 @@ export const BackButton = ({ text = 'Back', location }: BackButtonParams) => {
 
   return (
     <Button
-      className="margin-2 text-bold"
+      className="margin-top-2 text-bold"
       appearance="unstyled"
       text={
         <TextWithIcon
