+++ client/src/containers/Landing/Landing.scss
@@ -1,4 +1,4 @@
-.Home {
+.Landing {
   .usa-hero {
     background-image: url('~@ctoec/component-library/dist/assets/images/hero.svg');
     color: white;
