+++ client/src/index.scss
@@ -1,6 +1,6 @@
 @import './components';
 @import './containers/App/App.scss';
-@import './containers/Home/Home.scss';
+@import './containers/Landing/Landing.scss';
 @import './containers/EditRecord/EditRecord.scss';
 @import './containers/Roster/Roster.scss';
