+++ client/src/shared/payloads/BatchUpload.ts
@@ -1,5 +1,5 @@
 export interface BatchUpload {
   id?: number;
-  activeEnrollments: number;
-  withdrawnEnrollments: number;
+  active: number;
+  withdrawn: number;
 }
